Update Ejercicio 2.py

+++ Ejercicio 2.py
@@ -60,12 +60,12 @@ if __name__ == "__main__":
         contador = 1
         mensaje_solicitud ="Ingrese la Data"
 
-        Data_string = int(input("{} {}: ".format(mensaje_solicitud, contador)))
+        Data_string = float(input("{} {}: ".format(mensaje_solicitud, contador)))
 
     
         while Data_string != "*":
             contador += 1
-            numeros.append(int(Data_string))
+            numeros.append(float(Data_string))
             Data_string = input("{} {}: ".format(mensaje_solicitud, contador))
